Refactor log imports to use middleware/log for consistency across controllers and services
