+++ src/models/Log.ts
@@ -1,5 +1,5 @@
 import mongoose, { Document, Schema } from "mongoose";
-import { createTime } from "../services/logService";
+import { createTime } from "../middleware/log";
 
 // Interface for the Log document
 export interface ILog extends Document {
